requirements.txt
Fabric==1.2.2
South==0.7.3
amqplib==1.0.0
anyjson==0.3.1
celery==2.3.1
django-celery==2.3.3
django-debug-toolbar==0.8.5
django-picklefield==0.1.9
django-tastypie==0.9.9
epio==0.2.8
kombu==1.3.0
mimeparse==0.1.3
oauth2==1.5.170
ordereddict==1.1
paramiko==1.7.7.1
pycrypto==2.3
pyparsing==1.5.6
python-dateutil==1.5
python-digest==1.7
simplejson==2.1.6
twython==1.4.2
wsgiref==0.1.2
yolk==0.4.1
